version1/requirements.txt
# Main dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pillow>=10.0.0
matplotlib>=3.7.0
tqdm>=4.66.0
pyyaml>=6.0
scikit-image>=0.20.0
opencv-python>=4.8.0

# Diffusion model dependencies
diffusers>=0.21.0
transformers>=4.30.0
accelerate>=0.22.0
# huggingface-hub is needed for cache management
huggingface-hub>=0.16.0

# NOTE: xformers installation might fail on some platforms
# If xformers fails, comment it out and the model will still work (just slower)
# xformers>=0.0.20

# Medical imaging and evaluation dependencies
monai>=1.2.0
clean-fid>=0.1.35
torchmetrics>=1.0.0
medmnist>=3.0.0

# Optional logging
wandb>=0.15.0

# Memory management
psutil>=5.9.0 